+++ src/extension.ts
@@ -75,6 +75,50 @@ export function activate(context: vscode.ExtensionContext) {
 		
 
 	});
+	const disposableInject = vscode.commands.registerTextEditorCommand('go-spring.inject', (textEditor, edit) => {
+		// The code you place here will be executed every time your command is executed
+		// Display a message box to the user
+		
+		const doc = textEditor.document;
+        let selection: vscode.Selection | vscode.Range = textEditor.selection;
+ 
+        //获取选中区域
+        if (selection.isEmpty) {
+            const start = new vscode.Position(0, 0);
+            const end = new vscode.Position(doc.lineCount - 1, doc.lineAt(doc.lineCount - 1).text.length);
+            selection = new vscode.Range(start, end);
+        }
+        
+        let text = doc.getText(selection);
+
+		// vscode.window.showInformationMessage('text='+text);
+		let fullText=doc.getText();
+
+		if(fullText.includes("InjectBean(")){
+			let retText=process.injectBeanReplace(text,fullText);
+			textEditor.edit( (builder:any) => {
+				// let insertPosition=selection.end.translate(1);
+				const start = new vscode.Position(0, 0);
+				const end = new vscode.Position(doc.lineCount - 1, doc.lineAt(doc.lineCount - 1).text.length);
+				let section = new vscode.Range(start, end);
+				builder.replace(section, retText);
+			});
+		}else{
+			let code=process.injectBean(text,fullText);
+			if(code===''){
+				vscode.window.showInformationMessage('no change.');
+			}else{
+				textEditor.edit( (builder:any) => {
+					let insertPosition=selection.end.translate(1);
+					builder.insert(insertPosition, code);
+				});
+			}
+		}
+
+		
+		
+
+	});
 
 	const disposableCommonSetter = vscode.commands.registerTextEditorCommand('go-spring.var-setter', (textEditor, edit) => {
  
@@ -168,11 +212,46 @@ export function activate(context: vscode.ExtensionContext) {
 
 	});
 
+	const structToInterface = vscode.commands.registerTextEditorCommand('go-spring.interface', (textEditor, edit) => {
+ 
+		const doc = textEditor.document;
+        let selection: vscode.Selection | vscode.Range = textEditor.selection;
+		const position = textEditor.selection.active;
+
+        //获取选中区域
+        if (selection.isEmpty) {
+            // const start = new vscode.Position(0, 0);
+            // const end = new vscode.Position(doc.lineCount - 1, doc.lineAt(doc.lineCount - 1).text.length);
+            // selection = new vscode.Range(start, end);
+			selection = new vscode.Range(position.translate(-1), position);
+        }
+        
+        let text = doc.getText(selection);
+
+		// vscode.window.showInformationMessage('text='+text);
+		let fullText=doc.getText();
+ 
+		let result=process.convertToInterface(position,selection,text,fullText);
+		if(result.Err !==''  ){
+			vscode.window.showInformationMessage('Err:'+result.Err);
+		}else if(  result.Code===''){
+			vscode.window.showInformationMessage('no change.');
+		}else{
+			textEditor.edit( (builder:any) => {
+				let position=new vscode.Position(result.Line,0);
+				builder.insert(position, result.Code);
+			});
+		}
+		
+
+	});
+
 	context.subscriptions.push(disposableTab);
 	context.subscriptions.push(disposableSetter);
 	context.subscriptions.push(disposableCommonGetter);
 	context.subscriptions.push(disposableCommonSetter);
 	context.subscriptions.push(disposableName);
+	context.subscriptions.push(disposableInject);
 }
 
 // This method is called when your extension is deactivated